+++ app/src/main/java/info/nightscout/android/medtronic/message/CloseConnectionRequestMessage.java
@@ -5,6 +5,8 @@ import java.util.concurrent.TimeoutException;
 
 import info.nightscout.android.USB.UsbHidDriver;
 import info.nightscout.android.medtronic.MedtronicCnlSession;
+import info.nightscout.android.medtronic.exception.ChecksumException;
+import info.nightscout.android.medtronic.exception.EncryptionException;
 
 /**
  * Created by volker on 10.12.2016.
