+++ app/src/main/java/info/nightscout/android/medtronic/message/CloseConnectionResponseMessage.java
@@ -1,6 +1,8 @@
 package info.nightscout.android.medtronic.message;
 
 import info.nightscout.android.medtronic.MedtronicCnlSession;
+import info.nightscout.android.medtronic.exception.ChecksumException;
+import info.nightscout.android.medtronic.exception.EncryptionException;
 
 /**
  * Created by lgoedhart on 10/05/2016.
