+++ app/src/main/java/info/nightscout/android/medtronic/message/ContourNextLinkBinaryResponseMessage.java
@@ -1,5 +1,7 @@
 package info.nightscout.android.medtronic.message;
 
+import info.nightscout.android.medtronic.exception.ChecksumException;
+
 /**
  * Created by lgoedhart on 26/03/2016.
  */
