+++ app/src/main/java/info/nightscout/android/medtronic/message/ContourNextLinkCommandResponse.java
@@ -1,5 +1,7 @@
 package info.nightscout.android.medtronic.message;
 
+import info.nightscout.android.medtronic.exception.ChecksumException;
+
 /**
  * Created by volker on 10.12.2016.
  */
