+++ app/src/main/java/info/nightscout/android/medtronic/message/MedtronicPumpMessage.java
@@ -10,6 +10,5 @@ public class MedtronicPumpMessage extends ContourNextLinkMessage {
 
     protected MedtronicPumpMessage(MedtronicCnlSession pumpSession, byte[] bytes) {
         super(bytes);
-
     }
 }
