+++ app/src/main/java/info/nightscout/android/medtronic/message/MedtronicRequestMessage.java
@@ -10,6 +10,8 @@ import javax.crypto.spec.SecretKeySpec;
 
 import info.nightscout.android.USB.UsbHidDriver;
 import info.nightscout.android.medtronic.MedtronicCnlSession;
+import info.nightscout.android.medtronic.exception.ChecksumException;
+import info.nightscout.android.medtronic.exception.EncryptionException;
 
 /**
  * Created by lgoedhart on 26/03/2016.
