+++ app/src/main/java/info/nightscout/android/medtronic/message/PumpTimeResponseMessage.java
@@ -7,6 +7,8 @@ import java.nio.ByteOrder;
 import java.util.Date;
 
 import info.nightscout.android.medtronic.MedtronicCnlSession;
+import info.nightscout.android.medtronic.exception.ChecksumException;
+import info.nightscout.android.medtronic.exception.EncryptionException;
 
 /**
  * Created by lgoedhart on 27/03/2016.
